+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @navikt/sif-common-forms-ds
 
+## 9.0.7
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@1.11.3
+    -   @navikt/sif-common-core-ds@9.17.21
+    -   @navikt/sif-common-utils@3.46.10
+
+## 9.0.6
+
+### Patch Changes
+
+-   Updated dependencies [12cfdc9]
+    -   @navikt/sif-common-core-ds@9.17.20
+
 ## 9.0.5
 
 ### Patch Changes